fixes file exists check, improved tests and builds
