+++ README.md
@@ -9,7 +9,7 @@ storage adapters for S3 should work, but in practice they are hard to
 configure. This project uses the minio python client instead. Inspiration has
 been drawn from `django-s3-storage` and `django-storages`.
 
-It is tested on python 3 and django 1.9+ at the moment, but should be easily
+It is tested on python 3 and django 1.8 - 1.11, but should be easily
 portable to older versions. Python 2.7 is supported on a best effort basis,
 contributions are welcome. The current main blocker is issue #6.
 
@@ -17,13 +17,6 @@ The goal is to have a thoroughly tested, small codebase that delegates as
 much as possible to the minio client.
 
 Versioning is semver compliant.
-
-## Status
-
-This library is still a work in progress and not suitable for production
-use yet. Contributions and ideas are welcome (I mean it).
-
-See the issues to see what needs to be done for a 1.0 release to happen.
 
 ## Installation
 
@@ -50,6 +43,16 @@ does not already exist (default: `False`)
 - `MINIO_STORAGE_AUTO_CREATE_STATIC_BUCKET`: whether to create the bucket if it
 does not already exist (default: `False`)
 
+- `MINIO_PARTIAL_URL`: by default a `MINIO_STORAGE_ENDPOINT` is used as a base for file URLs.
+However when deployed in docker cluster, it is desired to return only the path and combine with `MINIO_PARTIAL_URL_BASE`
+ (default: `False`)
+- `MINIO_PARTIAL_URL_BASE`: base for the file's URL. It must be a valid URI without trailing slash at the end,
+e.g. 'http://example.com' or 'http://localhost:9000'
+
+- `MINIO_STORAGE_MEDIA_USE_PRESIGNED`: Determines if the media file URLs should be pre-signed (default: `False`)
+- `MINIO_STORAGE_STATIC_USE_PRESIGNED`: Determines if the static file URLs should be pre-signed (default: `False`)
+By default set to False.
+
 ## Short Example
 ```
 STATIC_URL = '/static/'
@@ -57,7 +60,7 @@ STATIC_ROOT = './static_files/'
 
 DEFAULT_FILE_STORAGE = "minio_storage.storage.MinioMediaStorage"
 STATICFILES_STORAGE = "minio_storage.storage.MinioStaticStorage"
-MINIO_STORAGE_ENDPOINT = 'localhost:9000'
+MINIO_STORAGE_ENDPOINT = 'minio:9000'
 MINIO_STORAGE_ACCESS_KEY = 'KBP6WXGPS387090EZMG8'
 MINIO_STORAGE_SECRET_KEY = 'DRjFXylyfMqn2zilAr33xORhaYz5r9e8r37XPz3A'
 MINIO_STORAGE_USE_HTTPS = False
@@ -65,6 +68,9 @@ MINIO_STORAGE_MEDIA_BUCKET_NAME = 'local-media'
 MINIO_STORAGE_AUTO_CREATE_MEDIA_BUCKET = True
 MINIO_STORAGE_STATIC_BUCKET_NAME = 'local-static'
 MINIO_STORAGE_AUTO_CREATE_STATIC_BUCKET = True
+
+MINIO_PARTIAL_URL = True
+MINIO_PARTIAL_URL_BASE = 'http://localhost:9000'
 ```
 
 ## Logging
