+++ dev-requirements.txt
@@ -1,5 +1,5 @@
 coverage
-minio
+minio==2.2.2
 mock
 requests
 
@@ -7,3 +7,4 @@ requests
 pytest==3.0.5
 pytest-django==3.1.2
 pytest-cov==2.4.0
+pytest-pythonpath
