+++ tests/test_app/tests/__init__.py
@@ -1,0 +1,4 @@
+from .bucket_tests import *
+from .delete_tests import *
+from .retrieve_tests import *
+from .upload_tests import *
