+++ tox.ini
@@ -1,31 +1,34 @@
 [pytest]
+python_paths = ./tests
+changedir = ./tests
+DJANGO_SETTINGS_MODULE = django_minio_storage_tests.settings
 addopts=--tb=short
-
-[tox]
-envlist =
-       {py27,py33,py34,py35}-django18,
-       {py27,py34,py35}-django{19,110},
-       {py27,py34,py35,py36}-django111
-
-[testenv]
-deps=pytest # install pytest in the venvs
-commands=py.test
+python_files = tests.py test_*.py *_tests.py
 
 [travis:env]
 DJANGO =
-    1.8: django18
-    1.9: django19
-    1.10: django110
-    1.11: django111
+        1.8: django18
+        1.9: django19
+        1.10: django110
+        1.11: django111
 
 [testenv]
-commands = ./runtests.py --fast {posargs} --coverage -rw
+commands = pytest
 setenv =
-       PYTHONDONTWRITEBYTECODE=1
-       PYTHONWARNINGS=once
+        PYTHONDONTWRITEBYTECODE=1
+        PYTHONWARNINGS=once
+        MINIO_STORAGE_ENDPOINT=localhost:9000
+        MINIO_STORAGE_ACCESS_KEY=weak_access_key
+        MINIO_STORAGE_SECRET_KEY=weak_secret_key
 deps =
         django18: Django>=1.8,<1.9
         django19: Django>=1.9,<1.10
         django110: Django>=1.10,<1.11
         django111: Django>=1.11,<2.0
-        -rrequirements/dev-requirements.txt
+        -rdev-requirements.txt
+
+[tox]
+envlist =
+       {py34,py35}-django18,
+       {py34,py35}-django{19,110},
+       {py34,py35}-django111
